fix(types): make generation deterministic across PHP versions
The name-sort getProperties override moves into a shared
SortsPropertiesByName trait and now covers ValueObjectTransformer too —
VO property order was still raw reflection order, so the first VO to
gain a trait or parent would have split the 8.4/8.5 CI legs. The base
profile also fails on a missing oxfmt binary instead of silently
writing unformatted output that diffs against the committed file.
generated.ts reorders VO props alphabetically (cosmetic).
