+++ src/Support/TypeScript/ComponentTransformer.php
@@ -29,6 +29,7 @@ use Spatie\TypeScriptTransformer\TypeScriptNodes\TypeScriptString;
 final class ComponentTransformer extends ClassTransformer
 {
     use AllowsListedClasses;
+    use SortsPropertiesByName;
 
     /**
      * @param  array<int, class-string>  $allowed
@@ -129,25 +130,5 @@ final class ComponentTransformer extends ClassTransformer
             new MarkerRewriteClassPropertyProcessor(Filter::class, NodeTypeReference::for(...)),
             new MixedToUnknownClassPropertyProcessor,
         ];
-    }
-
-    /**
-     * Sort by name so the generated output is deterministic across PHP versions:
-     * ReflectionClass::getProperties() reports inherited and trait properties in a
-     * different order on 8.4 vs 8.5.
-     *
-     * @return array<PhpPropertyNode>
-     */
-    #[\Override]
-    protected function getProperties(PhpClassNode $phpClassNode): array
-    {
-        $properties = parent::getProperties($phpClassNode);
-
-        usort(
-            $properties,
-            fn (PhpPropertyNode $a, PhpPropertyNode $b): int => $a->getName() <=> $b->getName(),
-        );
-
-        return $properties;
     }
 }
