+++ src/Support/TypeScript/OxfmtFormatter.php
@@ -9,11 +9,15 @@ use Symfony\Component\Process\Process;
 
 /**
  * Formats generated TypeScript with the host app's oxfmt binary when available,
- * falling back to the package checkout for local development. Missing binaries
- * are treated as a no-op so generation still produces valid TypeScript.
+ * falling back to the package checkout for local development. In consumer apps
+ * a missing binary is a no-op so generation still produces valid TypeScript;
+ * strict mode (the package's own base profile) fails instead, because an
+ * unformatted generated.ts would spuriously diff against the committed file.
  */
 final class OxfmtFormatter implements Formatter
 {
+    public function __construct(private readonly bool $strict = false) {}
+
     /**
      * @param  array<int, string>  $files
      */
@@ -26,6 +30,12 @@ final class OxfmtFormatter implements Formatter
         $binary = $this->resolveBinary();
 
         if ($binary === null) {
+            if ($this->strict) {
+                throw new \RuntimeException(
+                    'oxfmt binary not found (run `npm install`); refusing to write unformatted generated types.',
+                );
+            }
+
             return;
         }
